requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
dill==0.3.1.1
future==0.18.2
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.25.0
h5py==2.10.0
idna==2.8
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.1
promise==2.2.1
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
six==1.13.0
tensorboard==2.0.1
tensorflow-datasets==1.3.0
tensorflow-metadata==0.15.1
termcolor==1.1.0
tqdm==4.39.0
urllib3==1.25.7
virtualenv==16.7.8
vk==2.0.2
vk-api==11.6.1
Werkzeug==0.16.0
wrapt==1.11.2